// src/api/models/UserListSchema.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type UserListSchema = {
    userId?: string;
    email?: string;
    name?: string;
    userImageUrl?: string;
    company?: string;
    role?: string;
    delFlg?: number;
    createdAt?: string;
    updatedAt?: string;
};